+++ modules/uhttp/uhttpclient/client.go
@@ -25,14 +25,23 @@ import (
 	"time"
 
 	"go.uber.org/fx/auth"
+
+	"github.com/opentracing/opentracing-go"
 )
 
 // New creates an http.Client that includes 2 extra outbound middleware: tracing and auth
 // they are going to be applied in following order: tracing, auth, remaining outbound middleware
 // and only if all of them passed the request is going to be send.
 // Client is safe to use by multiple go routines, if global tracer is not changed.
-func New(info auth.CreateAuthInfo, middleware ...OutboundMiddleware) *http.Client {
-	defaultMiddleware := []OutboundMiddleware{tracingOutbound(), authenticationOutbound(info)}
+func New(tracer opentracing.Tracer, info auth.CreateAuthInfo, middleware ...OutboundMiddleware) *http.Client {
+	defaultMiddleware := make([]OutboundMiddleware, 0, 2+len(middleware))
+	if tracer != nil {
+		defaultMiddleware = append(defaultMiddleware, tracingOutbound(tracer))
+		if info != nil {
+			defaultMiddleware = append(defaultMiddleware, authenticationOutbound(info))
+		}
+	}
+
 	defaultMiddleware = append(defaultMiddleware, middleware...)
 	return &http.Client{
 		Transport: newExecutionChain(defaultMiddleware, http.DefaultTransport),