+++ modules/uhttp/uhttpclient/client_middleware_test.go
@@ -55,7 +55,7 @@ func TestExecutionChain(t *testing.T) {
 
 func TestExecutionChainOutboundMiddleware(t *testing.T) {
 	execChain := newExecutionChain(
-		[]OutboundMiddleware{tracingOutbound()}, nopTransport{},
+		[]OutboundMiddleware{tracingOutbound(opentracing.NoopTracer{})}, nopTransport{},
 	)
 	resp, err := execChain.RoundTrip(_req())
 	assert.NoError(t, err)
@@ -64,7 +64,7 @@ func TestExecutionChainOutboundMiddleware(t *testing.T) {
 
 func TestExecutionChainOutboundMiddlewareError(t *testing.T) {
 	execChain := newExecutionChain(
-		[]OutboundMiddleware{tracingOutbound()}, errTransport{},
+		[]OutboundMiddleware{tracingOutbound(opentracing.NoopTracer{})}, errTransport{},
 	)
 	resp, err := execChain.RoundTrip(_req().WithContext(context.Background()))
 	assert.Error(t, err)
@@ -77,9 +77,10 @@ func withOpentracingSetup(
 	registerFunc auth.RegisterFunc,
 	fn func(tracer opentracing.Tracer),
 ) {
-	tracer, closer, err := tracing.InitGlobalTracer(
+	tracer, closer, err := tracing.CreateTracer(
 		&jconfig.Configuration{}, "Test", zap.NewNop(), tally.NoopScope,
 	)
+
 	defer closer.Close()
 	assert.NotNil(t, closer)
 	require.NoError(t, err)
@@ -125,23 +126,22 @@ func getExecChainWithAuth(t *testing.T) executionChain {
 func TestOutboundMiddlewareWithTracerErrors(t *testing.T) {
 	testCases := map[string]OutboundMiddleware{
 		"auth":    authenticationOutbound(fakeAuthInfo{_testYaml}),
-		"tracing": tracingOutbound(),
+		"tracing": tracingOutbound(opentracing.NoopTracer{}),
 	}
 
 	for name, middleware := range testCases {
-		op := func(tracer opentracing.Tracer) {
+		op := func() {
 			tr := &shadowTracer{
-				tracer,
+				opentracing.NoopTracer{},
 				func(sm opentracing.SpanContext, format interface{}, carrier interface{}) error {
 					return errors.New("Very bad tracer")
 				},
 				nil,
 			}
-			opentracing.InitGlobalTracer(tr)
 
 			execChain := newExecutionChain(
-				[]OutboundMiddleware{middleware}, nopTransport{})
-			span := tracer.StartSpan("test_method")
+				[]OutboundMiddleware{middleware, tracingOutbound(tr)}, nopTransport{})
+			span := tr.Tracer.StartSpan("test_method")
 			span.SetBaggageItem(auth.ServiceAuth, "testService")
 			sp := &shadowSpan{span, tr}
 			tr.span = sp
@@ -150,11 +150,9 @@ func TestOutboundMiddlewareWithTracerErrors(t *testing.T) {
 
 			_, err := execChain.RoundTrip(_req().WithContext(ctx))
 			assert.EqualError(t, err, "Very bad tracer")
-
-			opentracing.InitGlobalTracer(tracer)
 		}
 
-		t.Run(name, func(t *testing.T) { withOpentracingSetup(t, nil, op) })
+		t.Run(name, func(t *testing.T) { op() })
 	}
 }