+++ CHANGELOG.md
@@ -25,6 +25,8 @@
   maps, builtin types and maps.
 * [Breaking] `package dig` has moved from `go.uber.org/fx/dig` to a new home
   at `go.uber.org/dig`.
+* [Breaking] Pass a tracer the `uhttp/uhttpclient` constructor explicitly, instead
+  of using a global tracer. This will allow to use http client in parallel tests.
 
 ## v1.0.0-beta2 (09 Mar 2017)